Refactor styled components into common styled components
