+++ client/src/Components/Common/Answers.js
@@ -1,0 +1,13 @@
+import styled from "styled-components"
+
+const Answers = styled.div.attrs({
+  className: "flex flex-column ph3",
+})`
+  label {
+    display: flex;
+    align-items: center;
+    width: 100%;
+  }
+`
+
+export { Answers }
