+++ client/src/Components/Common/Headings.js
@@ -1,0 +1,9 @@
+import styled from "styled-components"
+
+const Intro = styled.div.attrs({
+  className: "flex flex-column items-center ph3",
+})`
+  padding-top: 2rem;
+`
+
+export { Intro }
