+++ client/src/Components/Layouts/EditProfile/SupportDetails.js
@@ -2,13 +2,20 @@ import React, { Component } from "react"
 import axios from "axios"
 import swal from "sweetalert"
 
-import { Button, Answers, Intro } from "../Register/Register.style"
+// import styled components
+import { Intro } from "../../Common/Headings"
+import { Button } from "../../Common/Buttons"
+import { Answers } from "../../Common/Answers"
+
+// import common components
 import TextInput from "../../Common/Questions/TextInput"
 import RadioInput from "../../Common/Questions/RadioInput"
 import CheckboxInput from "../../Common/Questions/CheckboxInput"
 import FileUploadInput from "../../Common/Questions/FileUploadInput"
 import TextFieldInput from "../../Common/Questions/TextFieldInput"
 
+// get id from url
+// NOTE: this is until cookies are implemented
 const pathName = window.location.pathname
 const id = pathName.split("/")[3]
