+++ client/src/Components/Layouts/Register/Register.style.js
@@ -1,34 +1,5 @@
 import styled from "styled-components"
 
-const Intro = styled.div.attrs({
-  className: "flex flex-column items-center ph3",
-})`
-  padding-top: 2rem;
-`
-
-const Button = styled.button.attrs({
-  className: "mp-bg-white mp-b--primary-color",
-})`
-  border-style: solid;
-  border-width: 1px;
-  border-radius: 2rem;
-  cursor: pointer;
-  margin-top: 2rem;
-  margin-left: 0.5rem;
-  margin-right: 0.5rem;
-  width: 40vw;
-  height: 3rem;
-
-  :hover {
-    background-color: var(--primary);
-    color: var(--white);
-  }
-  :focus {
-    background-color: var(--darkPrimary);
-    color: var(--white);
-  }
-`
-
 const CircleLarge = styled.div.attrs({
   className: "mp-b--primary-color",
 })`
@@ -64,23 +35,4 @@ const RegisterStepTwoWrapper = styled.div.attrs({
   className: "flex flex-column ph3",
 })``
 
-const Answers = styled.div.attrs({
-  className: "flex flex-column ph3",
-})`
-  label {
-    display: flex;
-    align-items: center;
-    width: 100%;
-  }
-`
-
-export {
-  Intro,
-  Button,
-  CircleLarge,
-  CircleMed,
-  CirclesContainer,
-  CircleSm,
-  RegisterStepTwoWrapper,
-  Answers,
-}
+export { CircleLarge, CircleMed, CirclesContainer, CircleSm, RegisterStepTwoWrapper }
